Added get current user info endpoint

+++ FileStore.Api/UseCases/File/GetUserUsedSize/GetUserUsedSizeDto.cs
@@ -4,5 +4,5 @@ public class GetUserUsedSizeDto
 {
     public int MaxSize { get; set; }
     public int File { get; set; }
-    public int PrecentageFilledIn { get; set; }
+    public double PrecentageFilledIn { get; set; }
 }
